Add bulk refund tests
