+++ src/test/java/org/folio/rest/impl/accountactionchecks/AccountsActionChecksAPITests.java
@@ -1,30 +1,21 @@
 package org.folio.rest.impl.accountactionchecks;
 
-import static org.folio.rest.domain.Action.PAY;
-import static org.folio.rest.domain.Action.TRANSFER;
 import static org.folio.rest.utils.ResourceClients.buildAccountCheckPayClient;
-import static org.folio.rest.utils.ResourceClients.buildAccountCheckRefundClient;
 import static org.folio.rest.utils.ResourceClients.buildAccountCheckTransferClient;
 import static org.folio.rest.utils.ResourceClients.buildAccountCheckWaiveClient;
-import static org.folio.rest.utils.ResourceClients.feeFineActionsClient;
 
-import org.apache.http.HttpStatus;
-import org.folio.rest.jaxrs.model.Feefineaction;
 import org.folio.rest.utils.ResourceClient;
 import org.junit.Before;
 import org.junit.Test;
 
 public class AccountsActionChecksAPITests extends AccountsActionChecksAPITestsBase {
 
-  private static final double REQUESTED_AMOUNT = 1.23;
   private static final String ERROR_MESSAGE_MUST_BE_POSITIVE = "Amount must be positive";
   private static final String ERROR_MESSAGE_INVALID_AMOUNT = "Invalid amount entered";
 
   private ResourceClient accountsCheckPayClient;
   private ResourceClient accountsCheckWaiveClient;
   private ResourceClient accountsCheckTransferClient;
-  private ResourceClient accountsCheckRefundClient;
-
 
   @Before
   public void setUp() {
@@ -33,7 +24,6 @@ public class AccountsActionChecksAPITests extends AccountsActionChecksAPITestsBa
     accountsCheckPayClient = buildAccountCheckPayClient(firstAccount.getId());
     accountsCheckWaiveClient = buildAccountCheckWaiveClient(firstAccount.getId());
     accountsCheckTransferClient = buildAccountCheckTransferClient(firstAccount.getId());
-    accountsCheckRefundClient = buildAccountCheckRefundClient(firstAccount.getId());
   }
 
 
@@ -53,29 +43,6 @@ public class AccountsActionChecksAPITests extends AccountsActionChecksAPITestsBa
   }
 
   @Test
-  public void checkRefundAmountShouldBeAllowed() {
-    double expectedRemainingAmount = 0.77;
-
-    final Feefineaction feeFineAction = new Feefineaction()
-      .withAccountId(firstAccount.getId())
-      .withUserId(firstAccount.getUserId())
-      .withAmountAction((REQUESTED_AMOUNT + expectedRemainingAmount) / 2);
-
-    feeFineActionsClient()
-      .post(feeFineAction.withTypeAction(PAY.getPartialResult()))
-      .then()
-      .statusCode(HttpStatus.SC_CREATED);
-
-    feeFineActionsClient()
-      .post(feeFineAction.withTypeAction(TRANSFER.getPartialResult()))
-      .then()
-      .statusCode(HttpStatus.SC_CREATED);
-
-    actionShouldBeAllowed(false, accountsCheckRefundClient,
-      String.valueOf(expectedRemainingAmount));
-  }
-
-  @Test
   public void checkPayAmountShouldNotBeAllowedWithExceededAmount() {
     actionCheckAmountShouldNotBeAllowedWithExceededAmount(false, accountsCheckPayClient);
   }
@@ -88,11 +55,6 @@ public class AccountsActionChecksAPITests extends AccountsActionChecksAPITestsBa
   @Test
   public void checkTransferAmountShouldNotBeAllowedWithExceededAmount() {
     actionCheckAmountShouldNotBeAllowedWithExceededAmount(false, accountsCheckTransferClient);
-  }
-
-  @Test
-  public void checkRefundAmountShouldNotBeAllowedWithExceededAmount() {
-    actionCheckRefundAmountShouldNotBeAllowedWithExceededAmount(accountsCheckRefundClient);
   }
 
   @Test
@@ -110,12 +72,6 @@ public class AccountsActionChecksAPITests extends AccountsActionChecksAPITestsBa
   @Test
   public void checkTransferAmountShouldNotBeAllowedWithNegativeAmount() {
     actionShouldNotBeAllowed(false, accountsCheckTransferClient, "-5.0",
-      ERROR_MESSAGE_MUST_BE_POSITIVE);
-  }
-
-  @Test
-  public void checkRefundAmountShouldNotBeAllowedWithNegativeAmount() {
-    actionShouldNotBeAllowed(false, accountsCheckRefundClient, "-5.0",
       ERROR_MESSAGE_MUST_BE_POSITIVE);
   }
 
@@ -138,12 +94,6 @@ public class AccountsActionChecksAPITests extends AccountsActionChecksAPITestsBa
   }
 
   @Test
-  public void checkRefundAmountShouldNotBeAllowedWithZeroAmount() {
-    actionShouldNotBeAllowed(false, accountsCheckRefundClient, "0.0",
-      ERROR_MESSAGE_MUST_BE_POSITIVE);
-  }
-
-  @Test
   public void checkPayAmountShouldBeNumeric() {
     actionShouldNotBeAllowed(false, accountsCheckPayClient, "abc",
       ERROR_MESSAGE_INVALID_AMOUNT);
@@ -158,12 +108,6 @@ public class AccountsActionChecksAPITests extends AccountsActionChecksAPITestsBa
   @Test
   public void checkTransferAmountShouldBeNumeric() {
     actionShouldNotBeAllowed(false, accountsCheckTransferClient, "abc",
-      ERROR_MESSAGE_INVALID_AMOUNT);
-  }
-
-  @Test
-  public void checkRefundAmountShouldBeNumeric() {
-    actionShouldNotBeAllowed(false, accountsCheckRefundClient, "abc",
       ERROR_MESSAGE_INVALID_AMOUNT);
   }
 
@@ -184,13 +128,6 @@ public class AccountsActionChecksAPITests extends AccountsActionChecksAPITestsBa
     removeAllFromTable(ACCOUNTS_TABLE);
     actionCheckShouldNotFailForNonExistentAccount(false, accountsCheckTransferClient);
   }
-
-  @Test
-  public void checkRefundAmountShouldNotFailForNonExistentAccount() {
-    removeAllFromTable(ACCOUNTS_TABLE);
-    actionCheckShouldNotFailForNonExistentAccount(false, accountsCheckRefundClient);
-  }
-
 
   @Test
   public void checkPayAmountShouldNotBeAllowedForClosedAccount() {
